fix form management upload and process link unlink from modal (#1563)

+++ projects/valtimo/process-link/src/lib/components/process-link-modal/process-link-modal.component.ts
@@ -24,6 +24,7 @@ import {
 } from '../../services';
 import {take} from 'rxjs/operators';
 import {ConfigService} from '@valtimo/shared';
+import {ProcessLinkEditMode} from '../../models';
 
 @Component({
   standalone: false,
@@ -79,9 +80,17 @@ export class ProcessLinkModalComponent {
   }
 
   unlinkButtonClick(): void {
-    this.stateService.startSaving();
-
     this.stateService.selectedProcessLink$.pipe(take(1)).subscribe(selectedProcessLink => {
+      if (this.processLinkStateService.processLinkEditMode === ProcessLinkEditMode.EMIT_EVENTS) {
+        this.processLinkStateService.sendProcessLinkDeleteEvent({
+          activityId: selectedProcessLink.activityId,
+        });
+
+        return;
+      }
+
+      this.stateService.startSaving();
+
       this.processLinkService.deleteProcessLink(selectedProcessLink.id).subscribe(
         () => {
           this.stateService.closeModal();
